requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
Django==4.1.7
django-bootstrap4==22.3
django-ckeditor==6.5.1
django-filter==22.1
django-js-asset==2.0.0
django-widget-tweaks==1.4.12
psycopg2-binary==2.9.5
soupsieve==2.4
sqlparse==0.4.3
